+++ lib/cistern/client.ex
@@ -20,17 +20,20 @@ defmodule Cistern.Redis.Client do
     config =
       config
       |> filter_empty()
-      |> drop_pool_configs()
+      |> drop_internal_configs()
 
     @redis_module.start_link(config)
   end
 
   @doc false
   @impl true
-  def handle_call({command, args, opts}, _from, conn) do
-    Logger.debug("Execute Redix function `:#{command}` with Redis command #{inspect(args)}",
-      ansi_color: :green
-    )
+  def handle_call({command, args, opts}, _from, conn)
+      when command in [:command, :pipeline, :noreply_pipeline] do
+    if Application.get_env(:cistern, :debug_commands, false) do
+      Logger.debug("Execute Redix function `:#{command}` with Redis command #{inspect(args)}",
+        ansi_color: :green
+      )
+    end
 
     {:reply, apply(@redis_module, command, [conn, args, opts]), conn}
   end
@@ -41,7 +44,7 @@ defmodule Cistern.Redis.Client do
     end)
   end
 
-  defp drop_pool_configs(configs) do
-    Keyword.drop(configs, [:pool_size, :pool_max_overflow, :pool_timeout])
+  defp drop_internal_configs(configs) do
+    Keyword.drop(configs, [:pool_size, :pool_max_overflow, :pool_timeout, :coerce])
   end
 end